* commit '343568706cd090ea4d256cac050d2e4e7299376e': (30 commits)
  NTR - Fix code style
  SW-19930 - Fix snippets module pagesize
  SW-24440 - Fix wrong scope in cookie plugin which results in an error when switching the viewport
  SW-23620 - Update UPGRADE.md
  NTR - Allow container extensions
  SW-23620 - Countries with attributes fixes #2032
  SW-24423 - Fix mobile menu in content types
  SW-24428 - Fix custom http response codes in api
  SW-24430 - Fix double smarty block
  NTR - Fix commit hook for Enlight files
  SW-24400 - Updated UPGRADE.md
  SW-24400 - Fix PT-9148 fixes #2171
  NTR - Fix unsafe unserialize occurences
  NTR - Fix compiler pass LegacyApiResourcesPass
  SW-24432 - Fix snippet in blog module
  SW-24384 - Fix cart restoring when customer logouts
  Sw-24390 - Fixed snippets, window height and UPGRADE.md
  SW-24390 - Add separate option "stock" to "Apply Standard Data" fixes #2168
  SW-24376 - Reset filters on the detail page since we're using a global store
  SW-24420 - Prevent numeric field labels
  ...

+++ engine/Shopware/Components/Register/OptinCleanupService.php
@@ -57,15 +57,38 @@ class OptinCleanupService implements OptinCleanupServiceInterface
             $this->connection->beginTransaction();
             $queryBuilder = $this->connection->createQueryBuilder();
 
-            $queryBuilder->delete('s_core_optin')
+            $ids = $queryBuilder->select('optin.id')
+                ->from('s_core_optin', 'optin')
                 ->where('type != "swRegister"')
-                ->where('type LIKE "sw%"')
+                ->andWhere('type LIKE "sw%"')
                 ->andWhere('datum < NOW() - INTERVAL :interval DAY')
                 ->setParameter(':interval', $interval)
+                ->execute()
+                ->fetchAll(\PDO::FETCH_COLUMN);
+
+            if (!$ids) {
+                $this->connection->commit();
+
+                return false;
+            }
+
+            $queryBuilder = $this->connection->createQueryBuilder();
+            $queryBuilder->delete('s_core_optin')
+                ->where('id IN (:groupIds)')
+                ->setParameter(':groupIds', $ids, Connection::PARAM_INT_ARRAY)
+                ->execute();
+
+            $queryBuilder = $this->connection->createQueryBuilder();
+            $queryBuilder->update('s_user', 'user')
+                ->set('register_opt_in_id', 'NULL')
+                ->where('register_opt_in_id IN (:groupIds)')
+                ->setParameter(':groupIds', $ids, Connection::PARAM_INT_ARRAY)
                 ->execute();
 
             $this->connection->commit();
         } catch (DBALException $exp) {
+            $this->connection->rollBack();
+
             return false;
         }
